+++ .changeset/soft-worms-tan.md
@@ -1,0 +1,11 @@
+---
+"@rrweb/all": patch
+"rrdom-nodejs": patch
+"rrdom": patch
+"rrweb-snapshot": major
+"rrweb": patch
+"@rrweb/rrweb-plugin-canvas-webrtc-record": patch
+---
+
+`NodeType` enum was moved from rrweb-snapshot to @rrweb/types
+The following types where moved from rrweb-snapshot to @rrweb/types: `documentNode`, `documentTypeNode`, `legacyAttributes`, `textNode`, `cdataNode`, `commentNode`, `elementNode`, `serializedNode`, `serializedNodeWithId`, `serializedElementNodeWithId`, `serializedTextNodeWithId`, `IMirror`, `INode`, `mediaAttributes`, `attributes` and `DataURLOptions`
