+++ packages/plugins/rrweb-plugin-canvas-webrtc-record/src/index.ts
@@ -1,9 +1,9 @@
-import type { Mirror } from '@amplitude/rrweb-snapshot';
-import type {
-  ICrossOriginIframeMirror,
-  RecordPlugin,
-} from '@amplitude/rrweb-types';
 import SimplePeer from 'simple-peer-light';
+import type {
+  RecordPlugin,
+  ICrossOriginIframeMirror,
+  IMirror,
+} from '@amplitude/rrweb-types';
 import type { WebRTCDataChannel } from './types';
 
 export const PLUGIN_NAME = 'rrweb/canvas-webrtc@1';
@@ -28,7 +28,7 @@ export type CrossOriginIframeMessageEventContent = {
 
 export class RRWebPluginCanvasWebRTCRecord {
   private peer: SimplePeer.Instance | null = null;
-  private mirror: Mirror | undefined;
+  private mirror: IMirror<Node> | undefined;
   private crossOriginIframeMirror: ICrossOriginIframeMirror | undefined;
   private streamMap: Map<number, MediaStream> = new Map();
   private incomingStreams = new Set<MediaStream>();
