+++ packages/rrdom-nodejs/test/document-nodejs.test.ts
@@ -4,7 +4,7 @@
 import { describe, it, expect, beforeAll } from 'vitest';
 import * as fs from 'fs';
 import * as path from 'path';
-import { NodeType as RRNodeType } from '@amplitude/rrweb-snapshot';
+import { NodeType as RRNodeType } from '@amplitude/rrweb-types';
 import {
   RRCanvasElement,
   RRCDATASection,
