+++ packages/rrdom/test/virtual-dom.test.ts
@@ -6,18 +6,17 @@ import * as path from 'path';
 import * as puppeteer from 'puppeteer';
 import { vi } from 'vitest';
 import { JSDOM } from 'jsdom';
+import { buildNodeWithSN, Mirror } from '@amplitude/rrweb-snapshot';
 import {
-  buildNodeWithSN,
   cdataNode,
   commentNode,
   documentNode,
   documentTypeNode,
   elementNode,
-  Mirror,
   NodeType,
   NodeType as RRNodeType,
   textNode,
-} from '@amplitude/rrweb-snapshot';
+} from '@amplitude/rrweb-types';
 import {
   buildFromDom,
   buildFromNode,
