+++ packages/rrweb-snapshot/tsconfig.json
@@ -14,6 +14,9 @@
   },
   "references": [
     {
+      "path": "../types"
+    },
+    {
       "path": "../utils"
     }
   ]
