+++ packages/rrweb/src/record/iframe-manager.ts
@@ -1,13 +1,14 @@
-import type { Mirror, serializedNodeWithId } from '@amplitude/rrweb-snapshot';
-import { NodeType, genId } from '@amplitude/rrweb-snapshot';
+import type { Mirror } from '@amplitude/rrweb-snapshot';
+import { genId } from '@amplitude/rrweb-snapshot';
+import type { CrossOriginIframeMessageEvent } from '../types';
+import CrossOriginIframeMirror from './cross-origin-iframe-mirror';
+import { EventType, NodeType, IncrementalSource } from '@amplitude/rrweb-types';
 import type {
   eventWithTime,
   eventWithoutTime,
+  serializedNodeWithId,
   mutationCallBack,
 } from '@amplitude/rrweb-types';
-import { EventType, IncrementalSource } from '@amplitude/rrweb-types';
-import type { CrossOriginIframeMessageEvent } from '../types';
-import CrossOriginIframeMirror from './cross-origin-iframe-mirror';
 import type { StylesheetManager } from './stylesheet-manager';
 
 export class IframeManager {
