+++ packages/rrweb/src/replay/media/index.ts
@@ -1,12 +1,9 @@
-import {
-  type Emitter,
-  MediaInteractions,
-  ReplayerEvents,
-} from '@amplitude/rrweb-types';
+import type { Emitter } from '@amplitude/rrweb-types';
+import { MediaInteractions, ReplayerEvents } from '@amplitude/rrweb-types';
 import type { RRMediaElement } from '@amplitude/rrdom';
 import type { createPlayerService, createSpeedService } from '../machine';
-import type { Mirror, mediaAttributes } from '@amplitude/rrweb-snapshot';
-import type { mediaInteractionData } from '@amplitude/rrweb-types';
+import type { Mirror } from '@amplitude/rrweb-snapshot';
+import type { mediaInteractionData, mediaAttributes } from '@amplitude/rrweb-types';
 
 type MediaState = {
   isPlaying: boolean;
